+++ src/Recyclarr/Command/Initialization/IDefaultAppDataSetup.cs
@@ -1,0 +1,6 @@
+namespace Recyclarr.Command.Initialization;
+
+public interface IDefaultAppDataSetup
+{
+    void SetupDefaultPath(bool forceCreate = false);
+}