+++ src/Recyclarr/Migration/MigrationAutofacModule.cs
@@ -11,7 +11,11 @@ public class MigrationAutofacModule : Module
         builder.RegisterType<MigrationExecutor>().As<IMigrationExecutor>();
 
         // Migration Steps
-        builder.RegisterType<MigrateTrashYml>().As<IMigrationStep>();
-        builder.RegisterType<MigrateTrashUpdaterAppDataDir>().As<IMigrationStep>();
+        builder.RegisterTypes
+            (
+                typeof(MigrateTrashYml),
+                typeof(MigrateTrashUpdaterAppDataDir)
+            )
+            .As<IMigrationStep>();
     }
 }